Moving test helpers and auto-loading them
