+++ karma.conf.js
@@ -111,7 +111,8 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
-      'test/specs/**/*.spec.js'
+      'test/specs/__helpers.js',
+      'test/specs/**/*.spec.js',
     ],
 
 
@@ -124,6 +125,7 @@ module.exports = function(config) {
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
+      'test/specs/__helpers.js': ['webpack', 'sourcemap'],
       'test/specs/**/*.spec.js': ['webpack', 'sourcemap']
     },
