--- test/specs/__validateInvalidCharacterError.js
@@ -1,3 +1,0 @@
-module.exports = function validateInvalidCharacterError(error) {
-  expect(/character/i.test(error.message)).toEqual(true);
-};
