+++ test/specs/basicAuth.spec.js
@@ -1,5 +1,3 @@
-var setupBasicAuthTest = require('./__setupBasicAuthTest');
-
 describe('basicAuth without btoa polyfill', function () {
   setupBasicAuthTest();
 });
