+++ test/specs/helpers/btoa.spec.js
@@ -1,29 +1,26 @@
 var __btoa = require('../../../lib/helpers/btoa');
-var validateInvalidCharacterError = require('../__validateInvalidCharacterError');
 
 describe('btoa polyfill', function () {
   it('should behave the same as native window.btoa', function () {
+    // btoa doesn't exist in IE8/9
+    if (isOldIE && typeof Int8Array === 'undefined') {
+      return;
+    }
+
     var data = 'Hello, world';
     expect(__btoa(data)).toEqual(window.btoa(data));
   });
 
   it('should throw an error if char is out of range 0xFF', function () {
-    var err1, err2;
+    var err;
     var data = 'I ♡ Unicode!';
     
     try {
-      window.btoa(data);
-    } catch (e) {
-      err1 = e;
-    }
-
-    try {
       __btoa(data);
     } catch (e) {
-      err2 = e;
+      err = e;
     }
 
-    validateInvalidCharacterError(err1);
-    validateInvalidCharacterError(err2);
+    validateInvalidCharacterError(err);
   });
 });
